+++ compiler-core/build.gradle.kts
