+++ compiler-core/src/main/java/toothpick/compiler/common/ToothpickOptions.kt
@@ -32,8 +32,7 @@ data class ToothpickOptions(
     ),
     val crashWhenNoFactoryCanBeCreated: Boolean = false,
     val crashWhenInjectedMethodIsNotPackageVisible: Boolean = false,
-    val verboseLogging: Boolean = false,
-    val wrapScopeIntoWithControlFlow: Boolean = false
+    val verboseLogging: Boolean = false
 ) {
 
     companion object {
@@ -80,14 +79,6 @@ data class ToothpickOptions(
          * Might be useful for debugging or test purposes.
          */
         const val VerboseLogging = "toothpick_verbose_log"
-
-        /**
-         * The name of the annotation processor option to use `with` control flow
-         * to avoid `NAME_SHADOWING` issue.
-         *
-         * Might be useful for debugging or test purposes.
-         */
-        const val WrapScopeIntoWithControlFlow = "toothpick_wrap_scope_into_with_control_flow"
     }
 }
 
@@ -114,8 +105,5 @@ fun Map<String, String>.readOptions(): ToothpickOptions {
         verboseLogging = this[ToothpickOptions.VerboseLogging]
             ?.toBoolean()
             ?: default.verboseLogging,
-        wrapScopeIntoWithControlFlow = this[ToothpickOptions.WrapScopeIntoWithControlFlow]
-            ?.toBoolean()
-            ?: default.wrapScopeIntoWithControlFlow,
     )
 }
