+++ compiler-core/src/main/java/toothpick/compiler/common/generators/ClassNameExtensions.kt
