+++ compiler-core/src/main/java/toothpick/compiler/common/generators/targets/VariableInjectionTarget.kt
