+++ compiler-factory/src/main/java/toothpick/compiler/factory/FactoryProcessor.kt
@@ -310,11 +310,11 @@ class FactoryProcessor(env: SymbolProcessorEnvironment) : ToothpickProcessor(env
 
         val cannotCreateAFactoryMessage = (
             "Toothpick can't create a factory for it. " +
-            "If this class is itself a DI entry point (i.e. you call TP.inject(this) at some point), " +
-            "then you can remove this warning by adding @Suppress(\"Injectable\") to the class. " +
-            "A typical example is a class using injection to assign its fields, that calls TP.inject(this), " +
-            "but it needs a parameter for its constructor and this parameter is not injectable."
-        )
+                "If this class is itself a DI entry point (i.e. you call TP.inject(this) at some point), " +
+                "then you can remove this warning by adding @Suppress(\"Injectable\") to the class. " +
+                "A typical example is a class using injection to assign its fields, that calls TP.inject(this), " +
+                "but it needs a parameter for its constructor and this parameter is not injectable."
+            )
 
         val defaultConstructor = constructors.firstOrNull { constructor ->
             constructor.parameters.isEmpty()
