+++ compiler-factory/src/test/java/toothpick/compiler/factory/FactoryTest.kt
